Schema file renamed, validationHelpers util file created
